Move socs/agent/ to socs/common/
