+++ socs/agents/wiregrid_kikusui/agent.py
@@ -6,8 +6,8 @@ from ocs import ocs_agent, site_config
 from ocs.ocs_client import OCSClient
 from ocs.ocs_twisted import TimeoutLock
 
-from socs.agent import pmx
 from socs.agents.wiregrid_kikusui.drivers.common import openlog, writelog
+from socs.common import pmx
 
 
 class WiregridKikusuiAgent:
