+++ tests/common/test_pmx.py
@@ -2,7 +2,7 @@ import time
 
 import pytest
 
-from socs.agent.pmx import PMX, Command
+from socs.common.pmx import PMX, Command
 from socs.testing.device_emulator import create_device_emulator
 
 tcp_emulator = create_device_emulator({'ping': 'pong\r',
